refactor(cvvdp-gpu): u64::from(u32) instead of `u32 as u64` widening casts
Six lossless u32 to u64 widening casts switched to
u64::from(...). Sites:

  tests/common/mod.rs:409
    Some(q as u64)  ->  Some(u64::from(q))

  examples/time_12mp.rs:132 (per-pixel divisor):
    (W as u64) * (H as u64)  ->  u64::from(W) * u64::from(H)

  examples/time_size_sweep.rs:104 (per-bucket pixel count):
    (w as u64) * (h as u64)  ->  u64::from(w) * u64::from(h)

  benches/score.rs:59 and :129 (Throughput::Elements):
    (w * h) as u64           ->  u64::from(w) * u64::from(h)
    (W_256 * H_256) as u64   ->  u64::from(W_256) * u64::from(H_256)

clippy -W clippy::pedantic's cast_lossless flags:

  warning: casts from u32 to u64 can be expressed
           infallibly using From
    --> tests/common/mod.rs:409:68
    | Some(q as u64)
    |      ^^^^^^^^
    | help: an `as` cast can become silently lossy if the
    |       types change in the future

The u64::From<u32> impl encodes the widening intent explicitly
and would surface a hard compile error if a caller swapped u32
for a wider type. Throughput-math and the corpus-q comparison
evaluate to the same u64. Bench-site change also moves the
multiplication into u64 space, avoiding a hypothetical u32
overflow at large bench sizes (W * H currently fits u32 but
the new form is strictly safer).

Verification:
  cargo clippy -p cvvdp-gpu --features cuda --all-targets
    -W clippy::pedantic
       u32-to-u64 infallible-cast warnings:  0
  cargo test -p cvvdp-gpu --features cuda --test pipeline_score
    cvvdp_score_matches_v1_manifest
       1 passed; 0 failed

CHANGELOG entry under ### Changed (post-tick-238).

Tick 302.

+++ crates/cvvdp-gpu/benches/score.rs
@@ -56,7 +56,7 @@ fn bench_resolution(c: &mut Criterion, w: u32, h: u32, label: &str, include_host
 
     let group_name = format!("cvvdp_jod_{label}");
     let mut g = c.benchmark_group(&group_name);
-    g.throughput(Throughput::Elements((w * h) as u64));
+    g.throughput(Throughput::Elements(u64::from(w) * u64::from(h)));
 
     let client = Backend::client(&Default::default());
     let mut cvvdp = Cvvdp::<Backend>::new(client, w, h, CvvdpParams::PLACEHOLDER)
@@ -126,7 +126,7 @@ fn bench_at_quality(c: &mut Criterion, q: u32) {
 
     let group_name = format!("cvvdp_jod_256x256_q{q}");
     let mut g = c.benchmark_group(&group_name);
-    g.throughput(Throughput::Elements((W_256 * H_256) as u64));
+    g.throughput(Throughput::Elements(u64::from(W_256) * u64::from(H_256)));
 
     // All-host scalar path. Cvvdp::score used to route through this
     // before tick 213 switched to the GPU `compute_dkl_jod`; the
